+++ NEWS.md
@@ -32,6 +32,7 @@ Version 1.6.0.9000
 * [#543](https://github.com/tidyverse/lubridate/issues/543) Make `wday` work on character inputs as it is the case with all other day accessors.
 * [#566](https://github.com/tidyverse/lubridate/issues/566) Comparing durations and periods no-longer infloops.
 * [#556](https://github.com/tidyverse/lubridate/issues/556) Fix incorrect scoring of `y` format when it's the last in format order (as in `mdy`).
+* [#584](https://github.com/tidyverse/lubridate/issues/584) Fix interval by period division.
 * [#559](https://github.com/tidyverse/lubridate/issues/559) Parsing of alpha-months in English locales now drops correctly to low level C parsing. Thus, parsing with multiple orders containing `m` and `b` formats now works correctly.
 * [#570](https://github.com/tidyverse/lubridate/issues/570), [#574](https://github.com/tidyverse/lubridate/issues/574) Fix broken `date()` when called with missing argument.
 * [#567](https://github.com/tidyverse/lubridate/issues/567) Fix year update and rounding for leap years.